patch 9.1.0288: MS-Windows: compiler warning for size_t to int conversion
Problem:  MS-Windows: compiler warning for size_t to int conversion
          (after  v9.1.0282)
Solution: Use size_t instead of int in highlight_set_termgui_attr
          (Mike Williams)

closes: #14457

+++ src/highlight.c
@@ -860,7 +860,7 @@ highlight_reset_all(void)
 highlight_set_termgui_attr(int idx, char_u *key, char_u *arg, int init)
 {
     int		attr;
-    int		off;
+    size_t	off;
     keyvalue_T	target;
     keyvalue_T	*entry;
